tests/functional/requirements.txt
pytest==7.2.2
pytest-asyncio==0.21.0
pydantic==1.10.7
python-dotenv==0.21.1
aiohttp==3.8.4
allure-pytest==2.13.1
